Removed circular dependency in Home Builder and switched tiles to DS annotations
