+++ bundles/org.openhab.ui.basicui/OSGI-INF/.gitignore
@@ -1,0 +1,2 @@
+/*.xml
+
