+++ bundles/org.openhab.ui.classicui/OSGI-INF/.gitignore
@@ -1,0 +1,2 @@
+/*.xml
+
