+++ bundles/org.openhab.ui.homebuilder/OSGI-INF/.gitignore
@@ -1,0 +1,2 @@
+/*.xml
+
