--- bundles/org.openhab.ui.homebuilder/OSGI-INF/dashboardtile.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-	Copyright (c) 2015-2017 by the respective copyright holders.
-
-	All rights reserved. This program and the accompanying materials
-	are made available under the terms of the Eclipse Public License v1.0
-	which accompanies this distribution, and is available at
-	http://www.eclipse.org/legal/epl-v10.html
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" deactivate="deactivate" immediate="true" name="org.openhab.homebuilder">
-	<implementation class="org.openhab.ui.homebuilder.internal.HomeBuilderDashboardTile"/>
-	<reference bind="setHttpService" cardinality="1..1" interface="org.osgi.service.http.HttpService" name="HttpService" policy="static" unbind="unsetHttpService"/>
-	<service>
-		<provide interface="org.openhab.ui.dashboard.DashboardTile"/>
-	</service>
-</scr:component>
