+++ bundles/org.openhab.ui.homebuilder/src/main/java/org/openhab/ui/homebuilder/internal/HomeBuilderDashboardTile.java
@@ -9,46 +9,16 @@
 package org.openhab.ui.homebuilder.internal;
 
 import org.openhab.ui.dashboard.DashboardTile;
-import org.osgi.service.http.HttpService;
-import org.osgi.service.http.NamespaceException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.osgi.service.component.annotations.Component;
 
 /**
- * The dashboard tile and resource registering for Home Builder
+ * The dashboard tile for Home Builder
  *
  * @author Kuba Wolanin - Initial contribution
  *
  */
+@Component
 public class HomeBuilderDashboardTile implements DashboardTile {
-
-    public static final String HOMEBUILDER_ALIAS = "/homebuilder";
-
-    private final Logger logger = LoggerFactory.getLogger(HomeBuilderDashboardTile.class);
-
-    protected HttpService httpService;
-
-    protected void activate() {
-        try {
-            httpService.registerResources(HOMEBUILDER_ALIAS, "web", null);
-            logger.info("Started Home Builder at {}", HOMEBUILDER_ALIAS);
-        } catch (NamespaceException e) {
-            logger.error("Error during Home Builder startup: {}", e.getMessage());
-        }
-    }
-
-    protected void deactivate() {
-        httpService.unregister(HOMEBUILDER_ALIAS);
-        logger.info("Stopped Home Builder");
-    }
-
-    protected void setHttpService(HttpService httpService) {
-        this.httpService = httpService;
-    }
-
-    protected void unsetHttpService(HttpService httpService) {
-        this.httpService = null;
-    }
 
     @Override
     public String getName() {
