+++ bundles/org.openhab.ui.homebuilder/src/main/java/org/openhab/ui/homebuilder/internal/HomeBuilderServlet.java
@@ -1,0 +1,60 @@
+/**
+ * Copyright (c) 2015-2018 by the respective copyright holders.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+package org.openhab.ui.homebuilder.internal;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.http.HttpService;
+import org.osgi.service.http.NamespaceException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * The resource registering for Home Builder
+ *
+ * @author Kuba Wolanin - Initial contribution
+ *
+ */
+@Component(immediate = true)
+public class HomeBuilderServlet {
+
+    public static final String HOMEBUILDER_ALIAS = "/homebuilder";
+
+    private final Logger logger = LoggerFactory.getLogger(HomeBuilderServlet.class);
+
+    protected HttpService httpService;
+
+    @Activate
+    protected void activate() {
+        try {
+            httpService.registerResources(HOMEBUILDER_ALIAS, "web", null);
+            logger.info("Started Home Builder at {}", HOMEBUILDER_ALIAS);
+        } catch (NamespaceException e) {
+            logger.error("Error during Home Builder startup: {}", e.getMessage());
+        }
+    }
+
+    @Deactivate
+    protected void deactivate() {
+        httpService.unregister(HOMEBUILDER_ALIAS);
+        logger.info("Stopped Home Builder");
+    }
+
+    @Reference
+    protected void setHttpService(HttpService httpService) {
+        this.httpService = httpService;
+    }
+
+    protected void unsetHttpService(HttpService httpService) {
+        this.httpService = null;
+    }
+
+}
