+++ bundles/org.openhab.ui.paperui/OSGI-INF/.gitignore
@@ -1,0 +1,2 @@
+/*.xml
+
