--- bundles/org.openhab.ui.paperui/OSGI-INF/dashboardtile.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-	Copyright (c) 2015-2018 by the respective copyright holders.
-
-	All rights reserved. This program and the accompanying materials
-	are made available under the terms of the Eclipse Public License v1.0
-	which accompanies this distribution, and is available at
-	http://www.eclipse.org/legal/epl-v10.html
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.ui.paperui.dashboardtile">
-	<implementation class="org.openhab.ui.paperui.internal.PaperUIDashboardTile"/>
-	<service>
-		<provide interface="org.openhab.ui.dashboard.DashboardTile"/>
-	</service>
-</scr:component>
